Make gates first-class quantities in QVM
This commit does a good amount of restructuring to get it closer to
the ideal QAM implementation. Gates are no longer defined in the
source code as functions, but rather are included as a part of the
QVM. Since DEFGATE is still not supported, we do provide a set of
default gates with each QVM.

Gates are represented as subclasses of the abstract class GATE. Right
now, they're either parameterized or they're not (in which case
they're called "simple"). THe associated generic function
GATE-OPERATOR makes a matrix operator out of the gate, which is used
for execution.

The execution of the QVM used to be such that if an opcode was
encountered that wasn't known, the corresponding function name was
called. Now, if an unknown opcode is encountered, the QVM will look up
the opcode in its (privately) known gate set. If it does find it, then
it does a bit of work to parse out the instruction and apply the gate
accordingly.

The legacy behavior of calling a function is still useful, and is done
as a last resort.