optimization to MATRIX-MULTIPLY and amplitudes
This commit includes some mostly safe optimization for representing
the wavefunction as well as performing operations on it. This entails
using the newly introduced QUANTUM-STATE and QUANTUM-OPERATOR types.

The only function that has been intentionally optimized is
MATRIX-MULTIPLY, which has a SPEED=3 SAFETY=2 declaration.

Lastly, some functions that weren't needed were deleted, like the
in-place matrix multiplication as well as OUTER-PRODUCT.

The function KRONECKER-PRODUCT was moved to the test utilities since
it's not needed in the QVM.