create wavefunction.lisp; remove amplitude address; cleanup
This is a large commit that probably should have bit more granularly
split up. This commit basically adds no new functionality but cleans
things up a bit. The big changes are as follows.

1. Wavefunction module.

Functions that manipulate the wavefunction have been separated off
into wavefunction.lisp. This includes the venerable APPLY-OPERATOR
function.

Except for counting qubits, none of these actually depend on the QVM
itself.

2. AMPLITUDE-ADDRESS is removed.

Instead of maintaining the AMPLITUDE-ADDRESS register in the QVM, it
is now threaded through the computations of the wavefunction. This
makes multithreading easier and makes the QVM simpler. It does change
the API of a lot of functions.

3. RUN is now generic.

RUN is a generic function so it's possible to hook into it
programmatically.

4. General clean-up.

Some old functions dating back to the S-expression Quil days have been
removed. Things like CLASSICAL-CALL, and such.