abstract gate application
This commits structures the GATE classes a little bit more to include
room for gates that aren't defined by their matrix, but rather some
other computation. (Of course, every gate can in some way be turned
into a matrix.) This entailed adding a few more abstract classes,
namely STATIC-GATE and DYNAMIC-GATE, and making APPLY-GATE a generic
function.

The venerable APPLY-OPERATOR has been renamed to APPLY-MATRIX-OPERATOR
and is considered a low-level function for taking raw matrix objects
and applying them to raw wavefunction objects. This is used as a
subroutine to APPLY-GATE in, currently, all cases.

This is all to prepare for more efficient kinds of gate
representations and applications.