create APPLY-OPERATOR, now used to define APPLY-MATRIX-OPERATOR
APPLY-OPERATOR can modify the wavefunction with an arbitrary function,
instead of just a matrix. This will be useful for optimizing certain
transformations.

Making this change necessitated changing a lot of the matmul code. In
particular, result vectors are now allocated at the matmul level, as
opposed to the APPLY-*-OPERATOR level. The matmuls are now mutators on
a result vector created within MATRIX-MULTIPLY, and will mutate the
column vector being multiplied. This is backward incompatible with the
way things worked, but makes things a lot more generic at a higher
level.

This commit also does a few misc things, like export a few new
symbols.