Allow changing the number of threads in thread pool
This commit improves the behavior of prepare-for-parallelization.

The following are corrected:

* The message given by the assert call may be inaccurate when calling
  prepare-for-parallelization twice with different arguments.
* The doc string refers to the function by the wrong name.

Furthermore, calling prepare-for-parallelization once and then
calling it again with a different argument now changes the number
of threads. More precisely, the existing thread pool is freed and
a new one is created only if the number of threads requested differs
between the two calls.