Be explicit about allowed state transitions around WAIT-FUNCTION
Add new PQVM states WAITING, AWAKENING, and RUNNING, and use them to
determine valid state transitions around the wait/resume/run commands.

First, we only allow RUN-PROGRAM-ON-PERSISTENT-QVM if the PQVM is in
the state READY. Otherwise, we might run multiple programs that
contain WAIT instructions, making our lives miserable in the process.

Additionally, only wakeup from WAITing when the PQVM status has
changed from WAITING to something else. According to the
bordeaux-threads docs, it's possible for BT:CONDITION-WAIT to return
for reasons other than receiving a CONDITION-NOTIFY, so we check to
ensure that we really were woken up intentionally.