add support for serial kernels
The point of this PR is to allow smaller qubit simulations to be
faster. The QVM is pretty fast for large numbers of qubits, but
definitely doesn't do the best job for small numbers, which is an
important case to many people.

This PR adds macros for defining qubit application operators as
"stupid" loops that kick in when there aren't enough qubits for the
full-blown parallelized code.