Add HANDLE-POST-REQUEST-AND-CLEANUP
The garbage collection added at the end of HANDLE-POST-REQUEST in #199
prevents HANDLE-POST-REQUEST from returning a value. Some of the
handlers write to the response stream directly, but others just return
a value and let HUNCHENTOOT handle sending the response back to the
client.

This commit (1) ensures that a response value is always returned from
HANDLE-POST-REQUEST, and (2) ensures that cleanup / gc is always run,
even if the request signals an error.