WAIT!
Add a WAIT-FUNCTION slot to CLASSICAL-MEMORY-MIXIN that defaults to
WARNING-WAIT-FUNCTION, and call it in TRANSITION specialized on
((QVM CLASSICAL-MEMORY-MIXIN) (INSTR QUIL:WAIT)).

This was originally implemented as part of work on implementing WAIT
for persistent qvms, but factored out into a separate commit here.

This commit brought to you by the STYLEWARNING FOUNDATION FOR QVM
EMPOWERMENT AND CLASSICAL COMMUNITY SIMULATION.