[#227] fix redefine warning for compile-instruction generic
Here are two of the modules as they were ordered in qvm's asd file:

  - (:file "mixed-state-qvm")
  - (:file "compile-gate")

Mixed-state-qvm has a defmethod on compile-instruction. Compile-gate
has the defgeneric as well as several more defmethods. The Lisp system
was creating a standin defgeneric for mixed-state-qvm when it was
loaded first, and so then when it loaded the explicit defgeneric in
compile-gate, it issued the warning

  WARNING: redefining QVM::COMPILE-INSTRUCTION in DEFGENERIC

Solution: just switch the order, placing mixed-state-qvm after
compile-gate.